+++ src/OpenCVBuilder.ts
@@ -5,12 +5,10 @@ import log from 'npmlog';
 import { getLibsFactory } from './getLibsFactory.js';
 import { SetupOpencv } from './setupOpencv.js';
 import { Constant } from './constants.js';
-import { args2Option, genHelp, OpenCVBuildEnv, OpenCVBuildEnvParams, OPENCV_PATHS_ENV } from './BuildEnv.js';
+import OpenCVBuildEnv from './OpenCVBuildEnv.js';
+import { args2Option, genHelp, OpenCVBuildEnvParams, OPENCV_PATHS_ENV } from './misc.js';
 
-// import { fileURLToPath } from 'url';
-// import { dirname } from 'path';
 export class OpenCVBuilder {
-  // public readonly opencvVersion: string;
   public readonly constant: Constant;
   public readonly getLibs: getLibsFactory;
   public readonly env: OpenCVBuildEnv;