+++ src/constants.ts
@@ -1,27 +1,5 @@
 import { OpenCVBuilder } from "./OpenCVBuilder.js"
 
-const opencvModules = [
-  'core',
-  'highgui',
-  'imgcodecs',
-  'imgproc',
-  'features2d',
-  'calib3d',
-  'photo',
-  'objdetect',
-  'ml',
-  'video',
-  'videoio',
-  'videostab',
-  'dnn',
-  'face',
-  'text',
-  'tracking',
-  'xfeatures2d',
-  'ximgproc'
-] as const;
-
-export type opencvModulesType = typeof opencvModules[number] | 'world';
 
 export class Constant {
   constructor(private readonly builder: OpenCVBuilder) {
@@ -29,10 +7,11 @@ export class Constant {
 
   opencvRepoUrl = 'https://github.com/opencv/opencv.git'
   // opencvRepoUrl = 'c:/cache/opencv'
+
   opencvContribRepoUrl = 'https://github.com/opencv/opencv_contrib.git'
   // opencvContribRepoUrl = 'c:/cache/opencv_contrib'
-  
-  opencvModules = opencvModules;
+
+//   opencvModules = opencvModules;
 
   cmakeVsCompilers: { [version: string]: string } = {
     '10': 'Visual Studio 10 2010',
@@ -43,55 +22,5 @@ export class Constant {
     '16': 'Visual Studio 16 2019',
     '17': 'Visual Studio 17 2022',
   }
-
-  cmakeArchs: { [arch: string]: string } = {
-    'x64': ' Win64',
-    'ia32': '',
-    'arm': ' ARM'
-  }
-
-  defaultCmakeFlags = () => {
-    return [
-      `-DCMAKE_INSTALL_PREFIX=${this.builder.env.opencvBuild}`,
-      '-DCMAKE_BUILD_TYPE=Release',
-      '-DBUILD_EXAMPLES=OFF', // do not build opencv_contrib samples
-      '-DBUILD_DOCS=OFF',
-      '-DBUILD_TESTS=OFF',
-      '-DBUILD_PERF_TESTS=OFF',
-      '-DBUILD_JAVA=OFF',
-      '-DCUDA_NVCC_FLAGS=--expt-relaxed-constexpr',
-      '-DBUILD_opencv_apps=OFF',
-      '-DBUILD_opencv_aruco=OFF',
-      '-DBUILD_opencv_bgsegm=OFF',
-      '-DBUILD_opencv_bioinspired=OFF',
-      '-DBUILD_opencv_ccalib=OFF',
-      '-DBUILD_opencv_datasets=OFF',
-      '-DBUILD_opencv_dnn_objdetect=OFF',
-      '-DBUILD_opencv_dpm=OFF',
-      '-DBUILD_opencv_fuzzy=OFF',
-      '-DBUILD_opencv_hfs=OFF',
-      '-DBUILD_opencv_java_bindings_generator=OFF',
-      '-DBUILD_opencv_js=OFF',
-      '-DBUILD_opencv_img_hash=OFF',
-      '-DBUILD_opencv_line_descriptor=OFF',
-      '-DBUILD_opencv_optflow=OFF',
-      '-DBUILD_opencv_phase_unwrapping=OFF',
-      '-DBUILD_opencv_python3=OFF',
-      '-DBUILD_opencv_python_bindings_generator=OFF',
-      '-DBUILD_opencv_reg=OFF',
-      '-DBUILD_opencv_rgbd=OFF',
-      '-DBUILD_opencv_saliency=OFF',
-      '-DBUILD_opencv_shape=OFF',
-      '-DBUILD_opencv_stereo=OFF',
-      '-DBUILD_opencv_stitching=OFF',
-      '-DBUILD_opencv_structured_light=OFF',
-      '-DBUILD_opencv_superres=OFF',
-      '-DBUILD_opencv_surface_matching=OFF',
-      '-DBUILD_opencv_ts=OFF',
-      '-DBUILD_opencv_xobjdetect=OFF',
-      '-DBUILD_opencv_xphoto=OFF',
-      '-DBUILD_opencv_wechat_qrcode=OFF', // since 10/05/2022
-      '-DWITH_VTK=OFF'
-    ];
-  }
+  cmakeArchs: { [arch: string]: string } = { 'x64': ' Win64', 'ia32': '', 'arm': ' ARM' }
 }